+++ api/grpcserver/v2alpha1/mocks.go
@@ -14,8 +14,6 @@ import (
 	reflect "reflect"
 
 	types "github.com/spacemeshos/go-spacemesh/common/types"
-	identity "github.com/spacemeshos/go-spacemesh/identity"
-	builder "github.com/spacemeshos/go-spacemesh/sql/builder"
 	gomock "go.uber.org/mock/gomock"
 )
 
@@ -78,144 +76,6 @@ func (c *MockmalfeasanceInfoInfoCall) Do(f func(context.Context, types.NodeID) (
 
 // DoAndReturn rewrite *gomock.Call.DoAndReturn
 func (c *MockmalfeasanceInfoInfoCall) DoAndReturn(f func(context.Context, types.NodeID) (map[string]string, error)) *MockmalfeasanceInfoInfoCall {
-	c.Call = c.Call.DoAndReturn(f)
-	return c
-}
-
-// MockidentityState is a mock of identityState interface.
-type MockidentityState struct {
-	ctrl     *gomock.Controller
-	recorder *MockidentityStateMockRecorder
-	isgomock struct{}
-}
-
-// MockidentityStateMockRecorder is the mock recorder for MockidentityState.
-type MockidentityStateMockRecorder struct {
-	mock *MockidentityState
-}
-
-// NewMockidentityState creates a new mock instance.
-func NewMockidentityState(ctrl *gomock.Controller) *MockidentityState {
-	mock := &MockidentityState{ctrl: ctrl}
-	mock.recorder = &MockidentityStateMockRecorder{mock}
-	return mock
-}
-
-// EXPECT returns an object that allows the caller to indicate expected use.
-func (m *MockidentityState) EXPECT() *MockidentityStateMockRecorder {
-	return m.recorder
-}
-
-// All mocks base method.
-func (m *MockidentityState) All(ops builder.Operations) map[types.NodeID][]identity.StateInfo {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "All", ops)
-	ret0, _ := ret[0].(map[types.NodeID][]identity.StateInfo)
-	return ret0
-}
-
-// All indicates an expected call of All.
-func (mr *MockidentityStateMockRecorder) All(ops any) *MockidentityStateAllCall {
-	mr.mock.ctrl.T.Helper()
-	call := mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "All", reflect.TypeOf((*MockidentityState)(nil).All), ops)
-	return &MockidentityStateAllCall{Call: call}
-}
-
-// MockidentityStateAllCall wrap *gomock.Call
-type MockidentityStateAllCall struct {
-	*gomock.Call
-}
-
-// Return rewrite *gomock.Call.Return
-func (c *MockidentityStateAllCall) Return(arg0 map[types.NodeID][]identity.StateInfo) *MockidentityStateAllCall {
-	c.Call = c.Call.Return(arg0)
-	return c
-}
-
-// Do rewrite *gomock.Call.Do
-func (c *MockidentityStateAllCall) Do(f func(builder.Operations) map[types.NodeID][]identity.StateInfo) *MockidentityStateAllCall {
-	c.Call = c.Call.Do(f)
-	return c
-}
-
-// DoAndReturn rewrite *gomock.Call.DoAndReturn
-func (c *MockidentityStateAllCall) DoAndReturn(f func(builder.Operations) map[types.NodeID][]identity.StateInfo) *MockidentityStateAllCall {
-	c.Call = c.Call.DoAndReturn(f)
-	return c
-}
-
-// AllEligibilities mocks base method.
-func (m *MockidentityState) AllEligibilities() map[types.NodeID]map[types.LayerID][]types.VotingEligibility {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "AllEligibilities")
-	ret0, _ := ret[0].(map[types.NodeID]map[types.LayerID][]types.VotingEligibility)
-	return ret0
-}
-
-// AllEligibilities indicates an expected call of AllEligibilities.
-func (mr *MockidentityStateMockRecorder) AllEligibilities() *MockidentityStateAllEligibilitiesCall {
-	mr.mock.ctrl.T.Helper()
-	call := mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AllEligibilities", reflect.TypeOf((*MockidentityState)(nil).AllEligibilities))
-	return &MockidentityStateAllEligibilitiesCall{Call: call}
-}
-
-// MockidentityStateAllEligibilitiesCall wrap *gomock.Call
-type MockidentityStateAllEligibilitiesCall struct {
-	*gomock.Call
-}
-
-// Return rewrite *gomock.Call.Return
-func (c *MockidentityStateAllEligibilitiesCall) Return(arg0 map[types.NodeID]map[types.LayerID][]types.VotingEligibility) *MockidentityStateAllEligibilitiesCall {
-	c.Call = c.Call.Return(arg0)
-	return c
-}
-
-// Do rewrite *gomock.Call.Do
-func (c *MockidentityStateAllEligibilitiesCall) Do(f func() map[types.NodeID]map[types.LayerID][]types.VotingEligibility) *MockidentityStateAllEligibilitiesCall {
-	c.Call = c.Call.Do(f)
-	return c
-}
-
-// DoAndReturn rewrite *gomock.Call.DoAndReturn
-func (c *MockidentityStateAllEligibilitiesCall) DoAndReturn(f func() map[types.NodeID]map[types.LayerID][]types.VotingEligibility) *MockidentityStateAllEligibilitiesCall {
-	c.Call = c.Call.DoAndReturn(f)
-	return c
-}
-
-// AllProposals mocks base method.
-func (m *MockidentityState) AllProposals() map[types.NodeID][]*types.Proposal {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "AllProposals")
-	ret0, _ := ret[0].(map[types.NodeID][]*types.Proposal)
-	return ret0
-}
-
-// AllProposals indicates an expected call of AllProposals.
-func (mr *MockidentityStateMockRecorder) AllProposals() *MockidentityStateAllProposalsCall {
-	mr.mock.ctrl.T.Helper()
-	call := mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AllProposals", reflect.TypeOf((*MockidentityState)(nil).AllProposals))
-	return &MockidentityStateAllProposalsCall{Call: call}
-}
-
-// MockidentityStateAllProposalsCall wrap *gomock.Call
-type MockidentityStateAllProposalsCall struct {
-	*gomock.Call
-}
-
-// Return rewrite *gomock.Call.Return
-func (c *MockidentityStateAllProposalsCall) Return(arg0 map[types.NodeID][]*types.Proposal) *MockidentityStateAllProposalsCall {
-	c.Call = c.Call.Return(arg0)
-	return c
-}
-
-// Do rewrite *gomock.Call.Do
-func (c *MockidentityStateAllProposalsCall) Do(f func() map[types.NodeID][]*types.Proposal) *MockidentityStateAllProposalsCall {
-	c.Call = c.Call.Do(f)
-	return c
-}
-
-// DoAndReturn rewrite *gomock.Call.DoAndReturn
-func (c *MockidentityStateAllProposalsCall) DoAndReturn(f func() map[types.NodeID][]*types.Proposal) *MockidentityStateAllProposalsCall {
 	c.Call = c.Call.DoAndReturn(f)
 	return c
 }
